+++ packages/api-client-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @scalar/api-client-react
 
+## 1.3.107
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/api-client@2.31.3**
+  - [#8313](https://github.com/scalar/scalar/pull/8313): fix: correctly cache schema examples
+  - [#8254](https://github.com/scalar/scalar/pull/8254): fix: resolve OAuth2 URLs against interpolated server URL
+
+- **@scalar/types@0.6.10**
+
 ## 1.3.106
 
 ### Patch Changes