+++ packages/pre-post-request-scripts/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @scalar/scripts
 
+## 0.2.9
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/components@0.19.15**
+  - [#8319](https://github.com/scalar/scalar/pull/8319): fix(components): fix adjust sidebar section line height
+
+- **@scalar/oas-utils@0.8.3**
+  - [#8314](https://github.com/scalar/scalar/pull/8314): chore: limit concurrent operations while migrating workspaces
+
 ## 0.2.8
 
 ### Patch Changes